Allow binding DSL Rules to all items contained in a group (#4800)
This change allows writing rules that reference groups but execute based on the items within the group. This is primarily intended to extend the "triggeringItem" functionality provided in PR #4526, but also works with the other implicit variables.

Consider rules based on two switches contained in a group.

    Item SwitchGroup changed

will only execute when a switch change results in the group state changing. The script will not know which switch caused the change.

    Item Switch01 changed or
    Item Switch02 changed

and

    Member of SwitchGroup changed

are now functionally identical. The script will execute based on an individual switch state change without regard for the impact on the group.

The "Member of" version is advantageous because items can be added to or removed from the group without modifying the rule.

Signed-off-by: Jason Jones <jmsjones@gmail.com>